unify exiterror
